requirements.txt
django

django-allauth
django-filter
django-rest-auth

djangorestframework
markdown
django-filter

#dev
pylint

